+++ bapsf_motion/gui/calculators/__init__.py
@@ -1,6 +1,6 @@
 __all__ = ["LaPDXYZTransformCalculator", "LaPDXYZTransformCalculatorApp"]
 
-from bapsf_motion.gui.calculators.lapd_xyz_transform_calculator import (
+from bapsf_motion.gui.calculators.lapd_xyz_transform import (
     LaPDXYZTransformCalculator,
     LaPDXYZTransformCalculatorApp,
 )
